add hash-password command
