+++ server/src/command/add_admin.rs
